+ x8 + x16

+++ printfuncs/ft_putnbr.c
@@ -12,7 +12,7 @@
 
 #include "libft.h"
 
-void		ft_putnbr(long n)
+void		ft_putnbr(long long n)
 {
 	ft_putnbr_fd(n, 1);
 }
